Validation Set Example
Validation Set compares 3 linear regression models against the test set
to see how it performs on different model fits. This compares linear,
quadratic and cubic models. the Mean Squared Error obtained from the 3
models will be used as a benchmark. In this example it seems that after
fitting twice the models with different data, the Quadratic model seems
to be minimizing the MSE